Former-commit-id: 3f52dfd2f1957fd3411d301916a7d7db745d9ade

+++ helper_functions/opstack_metadata_utils.py
@@ -88,7 +88,7 @@ def generate_op_stack_chain_config_query_list(source_order=['oplabs', 'flipside'
                 return pd.DataFrame()  # Return an empty DataFrame if no data
 
 def gen_chain_ids_list_for_param(df_col):
-        chain_ids = df_col.astype(int).astype(str).tolist()
+        chain_ids = df_col.astype(str).tolist()
         chain_ids_string = ','.join(chain_ids)
         return chain_ids_string
 
@@ -97,7 +97,7 @@ def get_unique_chain_ids_from_dfs(dataframes):
         for df in dataframes:
                 if 'chain_id' in df.columns:
                         # Convert to float to handle decimal numbers, then to set to get unique values
-                        chain_ids = set(df['chain_id'].astype(int).astype(str).dropna())
+                        chain_ids = set(df['chain_id'].astype(str).dropna())
                         unique_chain_list.update(chain_ids)
                 else:
                         print(f"Warning: 'chain_id' column not found in one of the dataframes.")
